+++ Packages/com.frameworkbase.core/Tests/EditMode/SdkManagerTests.cs
@@ -124,6 +124,8 @@ namespace Framework.Tests
             public ISdkPurchaseService Purchase => null;
             public ISdkPushService Push => null;
             public ISdkPrivacyService Privacy => null;
+            public ISdkAdService Ad => null;
+            public ISdkComplianceService Compliance => null;
 
             public UniTask<SdkResult> InitializeAsync() => UniTask.FromResult(SdkResult.Ok());